added string library, removed strdup from code (not ISO C)

+++ src/diff/diff.c
@@ -1,6 +1,6 @@
 #include <diff/diff.h>
 
-#include <yadsl/posixstring.h>
+#include <string.h>
 #include <stddef.h>
 #include <stdlib.h>
 #include <math.h>
@@ -68,8 +68,8 @@ double yadsl_utils_diff(const char* s1, const char* s2)
 	double v[3], lv, cost = -1.0, * M, * N;
 	if (!s1 || !s2)
 		goto fail0;
-	l1 = strlen(s1) + 1;
-	l2 = strlen(s2) + 1;
+	l1 = strlen(s1) + (size_t) 1;
+	l2 = strlen(s2) + (size_t) 1;
 	if (l1 == 1 && l2 == 1)
 		return 0.0;
 	M = malloc(l1 * sizeof(double));
